Make image asset map part of project
