+++ fluid/Project.h
@@ -18,6 +18,7 @@
 #ifndef FLUID_PROJECT_H
 #define FLUID_PROJECT_H
 
+#include "app/Image_Asset_Map.h"
 #include "proj/undo.h"
 #include "proj/i18n.h"
 #include "nodes/Tree.h"
@@ -40,6 +41,11 @@ namespace app {
 class Project
 {
 public: // Member Variables
+  // Cache of image assets for this project. Must be declared before tree so
+  // that tree (and the Widget_Nodes it holds) is destroyed first, releasing
+  // all shared_ptr<Image_Asset> references before this map is torn down.
+  Image_Asset_Map image_assets;
+
   // Undo actions for this Project.
   proj::Undo undo { *this };
